+++ lib/src/configuration.dart
@@ -37,13 +37,19 @@ class Configuration {
 
   /// Creates a [Configuration] with schema objects for opening a [Realm].
   ///
+  /// [fifoFilesFallbackPath] enables FIFO special files.
   /// [readOnly] controls whether a [Realm] is opened as read-only.
   /// [inMemory] specifies if a [Realm] should be opened in-memory.
-  Configuration(List<SchemaObject> schemaObjects, {bool readOnly = false, bool inMemory = false})
+  Configuration(List<SchemaObject> schemaObjects, {String? fifoFilesFallbackPath, bool readOnly = false, bool inMemory = false})
       : _schema = RealmSchema(schemaObjects),
         _handle = realmCore.createConfig() {
     schemaVersion = 0;
     path = defaultPath;
+
+    if (fifoFilesFallbackPath != null) {
+      this.fifoFilesFallbackPath = fifoFilesFallbackPath;
+    }
+
     if (readOnly) {
       isReadOnly = true;
     }
@@ -109,6 +115,16 @@ class Configuration {
   /// When all in-memory instance of [Realm] is closed all data in that [Realm] is deleted.
   bool get isInMemory => realmCore.getConfigInMemory(this);
   set isInMemory(bool value) => realmCore.setConfigInMemory(this, value);
+
+  /// Gets or sets a value of FIFO special files location.
+  /// Opening a [Realm] creates a number of FIFO special files in order to
+  /// coordinate access to the [Realm] across threads and processes. If the [Realm] file is stored in a location
+  /// that does not allow the creation of FIFO special files (e.g. FAT32 filesystems), then the [Realm] cannot be opened.
+  /// In that case [Realm] needs a different location to store these files and this property defines that location.
+  /// The FIFO special files are very lightweight and the main [Realm] file will still be stored in the location defined
+  /// by the [path] you  property. This property is ignored if the directory defined by [path] allow FIFO special files.
+  String get fifoFilesFallbackPath => realmCore.getConfigFifoPath(this);
+  set fifoFilesFallbackPath(String value) => realmCore.setConfigFifoPath(this, value);
 }
 
 /// A collection of properties describing the underlying schema of a [RealmObject].