+++ lib/src/native/realm_core.dart
@@ -171,6 +171,16 @@ class _RealmCore {
     _realmLib.realm_config_set_in_memory(config.handle._pointer, value);
   }
 
+  String getConfigFifoPath(Configuration config) {
+    return _realmLib.realm_config_get_fifo_path(config.handle._pointer).cast<Utf8>().toDartString();
+  }
+
+  void setConfigFifoPath(Configuration config, String path) {
+    return using((Arena arena) {
+      _realmLib.realm_config_set_fifo_path(config.handle._pointer, path.toUtf8Ptr(arena));
+    });
+  }
+
   ConfigHandle createConfig() {
     final configPtr = _realmLib.realm_config_new();
     return ConfigHandle._(configPtr);